- added data validity to remove friend/enemy APIs

+++ SpyDuhApiProject2/Controllers/SpyDuhMembersController.cs
@@ -75,11 +75,15 @@ namespace SpyDuhApiProject2.Controllers
         [HttpPatch("removeFriend/{accountId}")]
         public IActionResult RemoveFriendFromSpyDuhAccount(Guid accountId, Guid friendId)
         {
+            var member = _spyDuhMembersRepository.GetById(accountId);
+            if (!(member.Friends.Any()) || !(member.Friends.Contains(friendId)))
+            {
+                return NotFound("No friends exist, or friend does not exist under this member.");
+            }
             _spyDuhMembersRepository.RemoveFriendFromSpyDuhAccount(accountId, friendId);
-            var updatedAccount = _spyDuhMembersRepository.GetById(accountId);
-            return Ok(updatedAccount);
+            return Ok(member);
         }
-
+        
         [HttpPatch("addEnemy/{accountId}")]
         public IActionResult AddEnemyToSpyDuhAccount(Guid accountId, Guid enemyId)
         {
@@ -91,9 +95,13 @@ namespace SpyDuhApiProject2.Controllers
         [HttpPatch("removeEnemy/{accountId}")]
         public IActionResult RemoveEnemyFromSpyDuhAccount(Guid accountId, Guid enemyId)
         {
+            var member = _spyDuhMembersRepository.GetById(accountId);
+            if (!(member.Enemies.Any()) || !(member.Enemies.Contains(enemyId)))
+            {
+                return NotFound("No enemies exist, or enemy does not exist under this member.");
+            }
             _spyDuhMembersRepository.RemoveEnemyFromSpyDuhAccount(accountId, enemyId);
-            var updatedAccount = _spyDuhMembersRepository.GetById(accountId);
-            return Ok(updatedAccount);
+            return Ok(member);
         }
 
         [HttpGet("enemies/{accountId}")]
